+++ src/download/mod.rs
@@ -4,15 +4,182 @@ use std::thread;
 
 use tokio::runtime::Runtime;
 
-use crate::download::dl_helpers::queue_libs;
 use crate::helpers;
-use crate::{
-	download::{dl_helpers::queue_assets, dl_types::DownloadEntry},
-	helpers::get_instance_folder,
-};
+use crate::helpers::get_instance_folder;
 
-mod dl_helpers;
-mod dl_types;
+use std::collections::HashMap;
+
+use serde::{Deserialize, Serialize};
+
+use std::vec;
+
+use reqwest::Client;
+
+pub fn download(
+	entry: &mut DownloadEntry,
+	prefix: PathBuf,
+) -> Result<Vec<u8>, Box<dyn std::error::Error>> {
+	let resp = reqwest::blocking::get(entry.url.clone())?;
+	let bytes = &resp.bytes()?;
+	helpers::write_bytes(
+		prefix
+			.join(entry.destination.clone())
+			.to_string_lossy()
+			.to_string(),
+		&bytes.clone(),
+	)?;
+
+	Ok(bytes.to_vec())
+}
+
+pub async fn get_version_manifest() -> Result<VersionManifest, Box<dyn std::error::Error>> {
+	let client = Client::new();
+	let raw_resp = client
+		.get("https://launchermeta.mojang.com/mc/game/version_manifest_v2.json")
+		.send()
+		.await?
+		.error_for_status()?
+		.text()
+		.await?;
+
+	let resp = serde_json::from_str::<VersionManifest>(&raw_resp).unwrap();
+
+	Ok(resp)
+}
+
+pub async fn get_version_json(url: String) -> Result<VersionJson, Box<dyn std::error::Error>> {
+	let client = Client::new();
+	let raw_resp = client
+		.get(&url)
+		.send()
+		.await?
+		.error_for_status()?
+		.text()
+		.await?;
+
+	let resp = serde_json::from_str::<VersionJson>(&raw_resp).unwrap();
+
+	Ok(resp)
+}
+
+pub async fn get_assets_vec(url: String) -> Result<MinecraftAssets, Box<dyn std::error::Error>> {
+	let client = Client::new();
+	let raw_resp = client
+		.get(&url)
+		.send()
+		.await?
+		.error_for_status()?
+		.text()
+		.await?;
+
+	let resp = serde_json::from_str::<MinecraftAssets>(&raw_resp).unwrap();
+
+	Ok(resp)
+}
+
+pub async fn queue_assets(
+	version: &VersionJson,
+) -> Result<Vec<DownloadEntry>, Box<dyn std::error::Error>> {
+	let assets: MinecraftAssets =
+		get_assets_vec(version.assetIndex["url"].as_str().unwrap().to_string()).await?;
+	let mut queue: Vec<DownloadEntry> = vec![];
+
+	for (k, a) in assets.objects.iter() {
+		queue.push(DownloadEntry {
+			url: format!(
+				"https://resources.download.minecraft.net/{}/{}",
+				&a.hash[0..2],
+				a.hash
+			),
+			destination: format!("assets/objects/{}/{}", &a.hash[0..2], a.hash),
+			size: Some(a.size),
+			sha1: None,
+			name: Some(k.to_string()),
+			executable: false,
+		});
+	}
+
+	Ok(queue)
+}
+
+pub async fn queue_libs(
+	version: &VersionJson,
+) -> Result<Vec<DownloadEntry>, Box<dyn std::error::Error>> {
+	let mut queue: Vec<DownloadEntry> = vec![];
+
+	for l in version.libraries.iter() {
+		// TODO: check for OS to make sure we aren't downloading
+		// anything we dont need
+		let dl = l["downloads"]["artifact"].clone();
+		queue.push(DownloadEntry {
+			size: Some(dl["size"].as_u64().unwrap() as usize),
+			destination: format!("libraries/{}", &dl["path"].as_str().unwrap()),
+			name: Some(dl["path"].as_str().unwrap().to_string()),
+			url: dl["url"].as_str().unwrap().to_string(),
+			sha1: Some(dl["sha1"].as_str().unwrap().to_string()),
+			executable: false,
+		})
+	}
+
+	Ok(queue)
+}
+
+#[derive(Debug, Clone)]
+pub struct DownloadEntry {
+	pub url: String,
+	pub destination: String,
+	pub size: Option<usize>,
+	pub sha1: Option<String>,
+	pub name: Option<String>,
+	pub executable: bool,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct VersionManifest {
+	pub latest: VersionManifestLatest,
+	pub versions: Vec<VersionManifestVersion>,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct VersionManifestLatest {
+	pub release: String,
+	pub snapshot: String,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct VersionManifestVersion {
+	pub id: String,
+	pub r#type: String,
+	pub url: String,
+	pub time: String,
+	#[allow(non_snake_case)]
+	pub releaseTime: String,
+	pub sha1: String,
+	#[allow(non_snake_case)]
+	pub complianceLevel: i8,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct VersionJson {
+	pub arguments: serde_json::Value,
+	#[allow(non_snake_case)]
+	pub assetIndex: serde_json::Value,
+	pub downloads: serde_json::Value,
+	pub libraries: Vec<serde_json::Value>,
+	#[serde(flatten)]
+	pub extra: serde_json::Value,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct MinecraftAsset {
+	pub hash: String,
+	pub size: usize,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct MinecraftAssets {
+	pub objects: HashMap<String, MinecraftAsset>,
+}
 
 pub fn new_instance(version: String, id: String, dl_threads: usize) {
 	let folder = get_instance_folder(&version);
@@ -26,9 +193,7 @@ pub fn install_minecraft(
 	threads: usize,
 ) -> Result<(), Box<dyn std::error::Error>> {
 	let rt = Runtime::new()?;
-	let versions = rt
-		.block_on(async { dl_helpers::get_version_manifest().await })
-		.unwrap();
+	let versions = rt.block_on(async { get_version_manifest().await }).unwrap();
 
 	let valid_version = match versions.versions.iter().find(|v| v.id == version) {
 		Some(v) => v,
@@ -36,7 +201,7 @@ pub fn install_minecraft(
 	};
 
 	let version_json = rt
-		.block_on(async { dl_helpers::get_version_json(valid_version.url.clone()).await })
+		.block_on(async { get_version_json(valid_version.url.clone()).await })
 		.unwrap();
 
 	let mut queue: Vec<DownloadEntry> = vec![];
@@ -89,7 +254,7 @@ pub fn install_minecraft(
 						e.size.map_or(0, |v| v)
 					);
 
-					if let Err(err) = dl_helpers::download(&mut e, get_instance_folder(&inst_id)) {
+					if let Err(err) = download(&mut e, get_instance_folder(&inst_id)) {
 						eprintln!(
 							"Failed to download {}: {}",
 							e.name.clone().unwrap_or("_".to_string()),
@@ -110,9 +275,7 @@ pub fn install_minecraft(
 
 pub fn list_versions() -> Result<(), Box<dyn std::error::Error>> {
 	let rt = Runtime::new()?;
-	let mut versions = rt
-		.block_on(async { dl_helpers::get_version_manifest().await })
-		.unwrap();
+	let mut versions = rt.block_on(async { get_version_manifest().await }).unwrap();
 
 	versions.versions.reverse();
